feat: working on domain modeling for tool type and inventory item
